Java crash when P2P service treats DHCP request
In July 2014 Google introduced a mechanism to let netd know
about routes on a p2p interface:
    ChgeId: I36effc438d5ac193a77174493bf196cb68a5b97a
    Title: Fix WiFi-Direct.
    Author: Sreeram Ramachandran <sreeram@google.com>
    Date: 2014-07-18 02:09:07

Unfortunately this mechanism doesn't handle all possible exceptions.
When device is connected as a STA and also as a P2P client and if
both DHCP leases ends at the same time, an IllegalArgumentException
can be thrown by method modifyRoute() in NetworkManagementService.
Such exception is not handled by the above new mechanism.

This patch fixes that!

Change-Id: I85b38784062c59b41c867395c499c43b1f8ffd24
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>